Started work on the multiblock nexus. No implementation yet, but the infrastructure is there. TBC when more research is done
This is a combination of 2 commits.
The first commit's message is:
Start of multiblock nexus

This is the 2nd commit message:
Small changes, but want to integrate a system that includes both immersive engineering type multiblocks as well as big reactor type multiblocks. I should also look into chisel multiblock handling for connected textures.